worked on dump summary feature

+++ packets/sourcepacket.cpp
@@ -78,7 +78,7 @@ SourcePacket* SourcePacket::makePacketFromData(unsigned char* pHeader_, unsigned
         dataFieldHeader->makeDataFieldHeaderFromData(pData_);
         // Move data pointer
 //        this->dataLength -= 2;
-        qulonglong id_ = (dataFieldHeader->getTimestampSeconds() << 32) + this->getSourceSequenceCount();
+        qulonglong id_ = ((qulonglong)dataFieldHeader->getTimestampSeconds() << 32) + this->getSourceSequenceCount();
         this->setId(id_);
     }
